requirements.txt
numpy == 1.26.4
pillow
pre-commit
black
